+++ boot/include/uart/uart.h
@@ -22,7 +22,7 @@
 
 #include <stdint.h>
 
-void uart_init(uint16_t baud_rate);
+void uart_init(uint32_t baud_rate);
 
 void uart_send(char c);
 void uart_sendstr(char *str);
